opt macos Dialog settings with NotNativeDialog and fix ci issues for all plugins

+++ app/db_tree/ecvPropertiesTreeDelegate.cpp
@@ -1144,7 +1144,7 @@ void ccPropertiesTreeDelegate::fillWithLineSet(
     addSeparator(tr("Line set"));
 
     appendRow(ITEM(tr("Points")),
-              ITEM(QLocale(QLocale::English).toString(_obj->points_.size())));
+              ITEM(QLocale(QLocale::English).toString(static_cast<qulonglong>(_obj->points_.size()))));
 
     appendRow(ITEM(tr("Segments")),
               ITEM(QLocale(QLocale::English).toString(_obj->segmentCount())));
